Handle broken pipes gracefully
When piping the output of vcdcat to programs closing the reading end
of the pipe before vcdcat finishes writing output the BrokenPipeError
exception is not handled. Given vcdcat leverages the builtin print
function throughout the implementation, we have installed the default
handler for the SIGPIPE signal taking special care to avoid breaking
compatibility with Windows (where SIGPIPE is not defined).